+++ lib/provider/detail_provider.dart
