+++ lib/ui/home.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:provider/provider.dart';
 import 'package:restaurant_app/data/api/api_services.dart';
-import 'package:restaurant_app/provider/restaurant_list_provider.dart';
-import 'package:restaurant_app/ui/restaurant_list.dart';
+import 'package:restaurant_app/provider/list_provider.dart';
+import 'package:restaurant_app/ui/list.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
