+++ lib/ui/list.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:restaurant_app/ui/restaurant_search.dart';
+import 'package:restaurant_app/ui/search.dart';
 import 'package:restaurant_app/widgets/search_field.dart';
 import 'package:restaurant_app/data/api/api_services.dart';
 import 'package:restaurant_app/data/model/restaurant.dart';
 import 'package:restaurant_app/widgets/card_restaurant.dart';
-import 'package:restaurant_app/provider/restaurant_list_provider.dart'
+import 'package:restaurant_app/provider/list_provider.dart'
     as restaurant_list_provider;
-import 'package:restaurant_app/provider/restaurant_search_provider.dart'
+import 'package:restaurant_app/provider/search_provider.dart'
     as restaurant_search_provider;
 
 class RestaurantListPage extends StatefulWidget {
