+++ lib/ui/review.dart
@@ -5,9 +5,9 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:http/http.dart' as http;
-import 'package:restaurant_app/ui/restaurant_list.dart';
+import 'package:restaurant_app/ui/list.dart';
 import 'package:restaurant_app/data/api/api_services.dart';
-import 'package:restaurant_app/provider/restaurant_list_provider.dart';
+import 'package:restaurant_app/provider/list_provider.dart';
 
 class ReviewPage extends StatefulWidget {
   final String id;
